refactor: extract shared text-source into a provider-neutral module
Greptile's round-N feedback: qwen-text-source re-exports from
mimo-text-source, which inverts the previous round's coupling
(qwen->openai) without removing it. Since the three providers are
distinct and could legitimately diverge (e.g., per-provider Chinese
normalisation or chunking rules), extract the canonical implementation
into shared-text-source.ts. mimo/qwen/openai now each re-export from
the shared module; any future per-provider tweak is a deliberate edit
in that one file rather than a silent inherit through a sibling.

Behaviour is identical — the three previous modules already had the
same implementation (modulo a docstring in mimo's version, preserved
on resolveReadingText in shared-text-source). Verified via ray lint
and ray build.
